requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
mysqlclient==2.2.4
passlib==1.7.4
python-jose==3.3.0
bcrypt==3.2.0